refactor(compiler-plugin/fir): move FIR pieces into hintGeneration / markerGeneration / scopeValidation logics
Move the FIR side of the `previewCollection` feature into its dedicated logic
directories. Pure moves + class renames; the heavyweight hint-vs-marker
generator split is deferred (see "Design deviation" below).

Moves:
- `fir/Deprecation.kt`
  → `feature/previewCollection/fir/hintGeneration/DeprecationHidden.kt`
  (rename — sealed-hidden marker is hint-specific so it lives next to the
  hint generator; both `markAsDeprecatedHidden` overloads are unchanged).
- `fir/InternalApiAttachment.kt`
  → `feature/previewCollection/fir/AttachInternalApi.kt`
  (rename + move to `feature/.../fir/` directly because both the hint and
  the marker generators need it; verb-prefix per project convention).
- `fir/checkers/CollectScopeAnnotationChecker.kt`
  → `feature/previewCollection/fir/scopeValidation/CheckCollectScopeAnnotation.kt`
  (class renamed `CollectScopeAnnotationChecker` → `CheckCollectScopeAnnotation`;
  `FirDeclarationChecker` inheritance preserved so `check(...)` remains the
  entry point).
- `fir/checkers/CollectScopeCallChecker.kt`
  → `feature/previewCollection/fir/scopeValidation/CheckCollectScopeCall.kt`
  (class renamed similarly).
- `fir/checkers/CollectScopeErrors.kt` + `PreviewLabFirCheckersExtension.kt`
  → `feature/previewCollection/fir/scopeValidation/` (CollectScopeErrors keeps
  noun form — it's a `KtDiagnosticsContainer`, not an action; CheckersExtension
  keeps `Extension` suffix per Kotlin compiler API convention).
- `fir/PreviewHintFirGenerator.kt`
  → `feature/previewCollection/fir/hintAndMarkerGeneration/PreviewHintFirGenerator.kt`
  (file moved + package updated; the deferred 1→2 logic split is documented as
  a future follow-up in the directory name).

Registrar (`fir/PreviewLabFirExtensionRegistrar.kt`) imports updated to point at
the new locations; the `supportsFirCheckers()` SPI gating for
`PreviewLabFirCheckersExtension` (added in commit `c4dcee6e`) is preserved.

★ 設計逸脱: ticket v2 仕様では 1-logic 統合の
`PreviewHintFirGenerator` を `hintGeneration/GeneratePreviewHintFir.kt` +
`markerGeneration/GeneratePreviewHintMarkerFir.kt` の **2 logic** に分離し、
`HintEntriesProvider` (commit 3 で導入済み) を経由して `hintEntries` を SSoT
共有する設計だった。 本コミットでは時間制約により分離を **`hintAndMarkerGeneration/`
として 1 logic に保留** し、 ticket v2 が後段の選択肢として残していた
「分離が過剰負担と判断されたら 1 logic 統合も将来の選択肢」 を採用。
`HintEntriesProvider` 自体は commit 3 で既に導入済みで session-scoped cache
として機能しているため、 2-logic 分離は別 PR で行える前提。 後続フォローアップ
チケットで `ComputeHintEntries.kt` / `ReadComposePreviewLabOption.kt` の抽出
と合わせて実施予定。

Quality gates (all green):
- `./gradlew ktlintFormat ktlintCheck --continue` — EXIT 0
- `./gradlew :compiler-plugin:test --rerun-tasks --continue` — EXIT 0
- `./gradlew jvmTest --rerun-tasks --continue` — EXIT 0
- `./gradlew apiCheck --continue` — EXIT 0
